requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-gis==0.12
Markdown==2.6.11
pbr==3.1.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
six==1.11.0
stevedore==1.28.0
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
